move the filter code
