+++ src/assets/api/unitsList.json
@@ -4610,7 +4610,7 @@
         "alt": "Gobelin"
       },
       {
-        "evolve": "assets/imgs/unitsList/Gobelin/evolves/chain_evolve_têtrouge.png",
+        "evolve": "assets/imgs/unitsList/Gobelin/evolves/chain_evolve_tetrouge.png",
         "alt": "Têtrouge"
       },
       {
